+++ editor/src/com/talosvfx/talos/editor/addons/scene/apps/spriteeditor/EditPanel.java
@@ -1,4 +1,4 @@
-package com.talosvfx.talos.editor.addons.scene.apps;
+package com.talosvfx.talos.editor.addons.scene.apps.spriteeditor;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
